Python/requirements.txt
requests
boto3
python-dotenv
 
